Add CSS support
In the development mode we let Webpack bundle the CSS code with JavaScript and
for "demo" and production targets, a separate stylesheet file gets created.

ES5 shim and sham as well as HTML5 shiv have been added to make sure demo and
production versions of the application can be run in IE8. Please note, it is
not currently possible to do it for the devevelopment version, partly because
Hot Module Loader breaks IE8.